requirements.txt
annotated-types==0.7.0
anyio==4.10.0
blinker==1.9.0
blis==1.3.0
cachetools==5.5.2
catalogue==2.0.10
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
cloudpathlib==0.22.0
colorama==0.4.6
confection==0.1.5
cryptography==45.0.7
cymem==2.0.11
dicttoxml==1.7.16
et_xmlfile==2.0.0
Flask==3.1.2
google-api-core==2.25.1
google-auth==2.40.3
google-cloud-language==2.17.2
google-genai==1.32.0
googleapis-common-protos==1.70.0
grpcio==1.74.0
grpcio-status==1.74.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
langcodes==3.5.0
language_data==1.3.0
marisa-trie==1.3.1
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
murmurhash==1.0.13
numpy==2.3.2
openpyxl==3.1.5
packaging==25.0
pandas==2.3.2
preshed==3.0.10
proto-plus==1.26.1
protobuf==6.32.0
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.5
rich==14.1.0
rsa==4.9.1
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
smart_open==7.3.0.post1
sniffio==1.3.1
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
tenacity==9.1.2
thinc==8.3.6
tqdm==4.67.1
typer==0.17.3
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
wasabi==1.1.3
weasel==0.4.1
websockets==15.0.1
Werkzeug==3.1.3
wrapt==1.17.3
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl
https://github.com/explosion/spacy-models/releases/download/fr_core_news_sm-3.8.0/fr_core_news_sm-3.8.0-py3-none-any.whl

